Five layouts, lazyloading and search improvement
Added four more layouts.
- Posters
- Posters with title
- Posters with movie information
- List
- List with movie information

All posters are now loaded with lazyloading

You can now click on director/writer/producer/music/cast inside movie
information, and it will find every movie they have worked with. It can
sometimes find more movies due to the fact it also searches in plot
outline, and it separates all names with space. So it can for an example
search for Will OR Smith. And find the movie Pirates of the Caribbean:
The Curse of the Black Pearl (Will Turner/Blacksmith)